[542494] toString has size_t in its second parameter
The size_t is adopted as type of the bufferSize parameter of toString
functions

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=542494

+++ src/core/datatypes/forte_any_string.h
@@ -127,7 +127,9 @@ class CIEC_ANY_STRING : public CIEC_ANY_ELEMENTARY{
      *   \return number of bytes used in the buffer
      *           -1 on error
      */
-    virtual int toUTF8(char*, unsigned int, bool) const{return 0;}
+    virtual int toUTF8(char*, size_t, bool) const {
+      return 0;
+    }
 #endif
 
   protected:
