+++ src/core/datatypes/forte_date_and_time.cpp
@@ -105,14 +105,14 @@ int CIEC_DATE_AND_TIME::fromString(const char *paValue){
   return nRetVal;
 }
 
-int CIEC_DATE_AND_TIME::toString(char* paValue, unsigned int paBufferSize) const{
+int CIEC_DATE_AND_TIME::toString(char* paValue, size_t paBufferSize) const {
   int nRetVal = -1;
 
   struct tm *ptm = getTimeStruct();
 
   if(0 != ptm){
     nRetVal = forte_snprintf(paValue, paBufferSize, "%04d-%02d-%02d-%02d:%02d:%02d.%03u", 1900 + ptm->tm_year, ptm->tm_mon + 1, ptm->tm_mday, ptm->tm_hour, ptm->tm_min, ptm->tm_sec, getMilliSeconds());
-    if((nRetVal < -1) || (nRetVal >= (int) paBufferSize)){
+    if((nRetVal < -1) || (nRetVal >= static_cast<int>(paBufferSize))) {
       nRetVal = -1;
     }
   }
